+++ CHANGELOG.md
@@ -19,6 +19,7 @@ also non string parameters (e.g. `Enum.join([1, 2], ",")`
 - Add support to Elixir `Enumerable` protocol also for `Enum.all?`, `Enum.any?`, `Enum.each`,
 `Enum.filter`, `Enum.flat_map`, `Enum.reject`, `Enum.chunk_by` and `Enum.chunk_while`
 - Support for `maps:merge_with/3`
+- Support for `lists:last/1` and `lists:mapfoldl/3`
 
 ### Changed